[FIX] long -> int 변환

+++ src/main/java/com/ssafy/sgdc/competition/MatchingRepo.java
@@ -1,11 +1,10 @@
 package com.ssafy.sgdc.competition;
 
-import com.ssafy.sgdc.domain.entity.Matching;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface MatchingRepo extends JpaRepository<Matching, Long> {
+public interface MatchingRepo extends JpaRepository<Matching, Integer> {
